requirements.txt
aiohttp
dnspython
rich
typer
ipwhois
tldextract
beautifulsoup4
pytest
pytest-asyncio
